workers/requirements_ai.txt
pika
opencv-python-headless
numpy
psycopg2-binary
scikit-learn==1.0.2
pandas==1.3.5